// src/models/interfaces/jaridata.ts
export interface jaridata {
    Id: string;
    daerah_irigasi: string;
    bendung: string;
    jaringan_irigasi:string;
    jaringan_irigasi_bangunan:string;
    saluran_primer:string;
    drain_inlet:string;
    saluran_sekunder:string;
    kondisi:string;
    x: string;
    y: string;
    foto: string;
}